[RAND-1]: dmeze: Added prisma cache strategy, Add Sharp package, fix bot
https://jira-eng-rtp3.cisco.com/jira/browse/RAND-1
