+++ package.json
@@ -4,13 +4,14 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "npx prisma generate --no-engine next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
     "@floating-ui/react": "^0.26.16",
-    "@prisma/client": "^5.13.0",
+    "@prisma/client": "^5.15.0",
+    "@prisma/extension-accelerate": "^1.1.0",
     "@react-spring/web": "^9.7.3",
     "@reduxjs/toolkit": "^2.1.0",
     "@stripe/react-stripe-js": "^2.7.0",
@@ -32,6 +33,7 @@
     "react-slick": "^0.30.1",
     "react-toastify": "^10.0.5",
     "react-wavify": "^1.10.0",
+    "sharp": "^0.33.4",
     "slick-carousel": "^1.8.1",
     "stripe": "^15.4.0",
     "uuid": "^9.0.1",
@@ -64,7 +66,7 @@
     "eslint-plugin-tailwindcss": "^3.14.0",
     "eslint-plugin-unused-imports": "^3.0.0",
     "postcss": "^8",
-    "prisma": "^5.13.0",
+    "prisma": "^5.15.0",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
