+++ scripts/build.ts
@@ -1,0 +1,3 @@
+import { bot } from '@/bot'
+
+bot.api.setWebhook(`${process.env.VERCEL_URL}/api/bot`)
